Cast advisory lock ID to bigint in scheduler
Previously the scheduler used a hardcoded integer literal for Postgres advisory locks in Prisma raw queries, which can cause bigint/type mismatch errors. This change explicitly casts the lock ID to bigint (1337::bigint) for both pg_try_advisory_lock and pg_release_advisory_lock and updates the comment accordingly to prevent runtime type issues when acquiring/releasing the distributed lock.

+++ apps/worker/src/matchmaking/scheduler.ts
@@ -35,8 +35,8 @@ const SCHEDULER_LOCK_ID = 1337;
 export async function scheduleMatches(): Promise<number> {
   // 0. Acquire Distributed Lock (Postgres Advisory Lock)
   // This ensures ONLY ONE worker replica runs the scheduler at a time.
-  // Using hardcoded ID to avoid bigint casting issues in Prisma raw queries.
-  const lock = await prisma.$queryRaw<[{ pg_try_advisory_lock: boolean }]>`SELECT pg_try_advisory_lock(1337)`;
+  // Using hardcoded ID with explicit bigint cast to avoid type mismatch errors.
+  const lock = await prisma.$queryRaw<[{ pg_try_advisory_lock: boolean }]>`SELECT pg_try_advisory_lock(1337::bigint)`;
   
   if (!lock || !lock[0].pg_try_advisory_lock) {
     return 0; // Lock is held by another worker replica
@@ -162,6 +162,6 @@ export async function scheduleMatches(): Promise<number> {
   return scheduled;
 } finally {
   // Always release the lock so other workers can try again in the next 30s cycle
-  await prisma.$executeRaw`SELECT pg_release_advisory_lock(1337)`;
+  await prisma.$executeRaw`SELECT pg_release_advisory_lock(1337::bigint)`;
 }
 }
